Get rid of "app" posts
On Facebook I find these "app" posts extremely annoying. For as many as you
knock down, some new app comes up and you have to go and block it. You are left
with a block list littered with apps, and in my case, people unfriended. This
commit removes ALL app posts on the feed, with no need to add apps to a block
list.